fix: fix based on pr comments
