+++ psycop/common/sequence_models/train.py
@@ -7,7 +7,9 @@ from pathlib import Path
 import lightning.pytorch as pl
 from torch.utils.data import DataLoader
 
-from .config_utils import flatten_nested_dict, load_config, parse_config
+from psycop.common.global_utils.config_utils import flatten_nested_dict
+
+from .config_utils import load_config, parse_config
 
 std_logger = logging.getLogger(__name__)
 
@@ -19,15 +21,6 @@ def populate_registry() -> None:
     It is also possible to do this using hooks, but this is more explicit
     and easier to debug for people who are not familiar with python setup hooks.
     """
-    from .embedders.BEHRT_embedders import create_behrt_embedder
-    from .logger import create_wandb_logger
-    from .model_layers import create_encoder_layer, create_transformers_encoder
-    from .optimizers import (
-        create_adam,
-        create_adamw,
-        create_linear_schedule_with_warmup,
-    )
-    from .tasks import create_behrt, create_encoder_for_clf
 
 
 populate_registry()
