+++ psycop/projects/sequence_models/fine_tune_w_config.py
@@ -1,5 +1,6 @@
 import datetime as dt
 from pathlib import Path
+from typing import Literal
 
 from psycop.common.feature_generation.loaders.raw.load_ids import SplitName
 from psycop.common.feature_generation.sequences.cohort_definer_to_prediction_times import (
@@ -9,10 +10,10 @@ from psycop.common.feature_generation.sequences.patient_loaders import (
     DiagnosisLoader,
     PatientLoader,
 )
-from psycop.common.sequence_models import train
 from psycop.common.sequence_models.dataset import PatientSlicesWithLabels
 from psycop.common.sequence_models.registry import Registry
 from psycop.common.sequence_models.tasks import BEHRTForMaskedLM
+from psycop.common.sequence_models.train import train
 from psycop.projects.t2d.feature_generation.cohort_definition.t2d_cohort_definer import (
     T2DCohortDefiner,
 )
@@ -28,7 +29,7 @@ def load_model_from_checkpoint(
 @Registry.datasets.register("patient_slices_with_labels_for_t2d")
 def create_patient_slices_with_labels_for_t2d(
     min_n_visits: int,
-    split_name: str,
+    split_name: Literal["train", "val", "test"],
     lookbehind_days: int = 365,
     lookahead_days: int = 365,
 ) -> PatientSlicesWithLabels:
